0.04

+++ core/src/com/MVlab/BrickBreaker/gameworld/GameRenderer.java
@@ -57,19 +57,13 @@ public class GameRenderer {
         physicWorld.step(Gdx.graphics.getDeltaTime(), 1, 1);
 
         //deleting bodies+++
-        ArrayList<Integer> deleteIndexes = new ArrayList<Integer>();
-        int index = -1;
         for (Brick brick : bricks) {
-            index++;
-            if (brick != null && brick.getHealth() <= 0) {
+            if (brick != null && brick.getHealth() <= 0 && brick.existing()) {
                 brick.getBody().setActive(false);
                 physicWorld.destroyBody(brick.getBody());
-                deleteIndexes.add(index);
+                brick.delete();
             }
         }
-        for (Integer delIndex: deleteIndexes)
-            bricks.remove(delIndex.intValue());
-        //deleting bodies---
 
         Gdx.gl.glClearColor(0, 0, 0, 1);
         Gdx.gl.glClear(GL20.GL_COLOR_BUFFER_BIT);
